+++ src/drivers/cmakeServerDriver.ts
@@ -26,6 +26,7 @@ import * as nls from 'vscode-nls';
 import { BuildPreset, ConfigurePreset, TestPreset } from '@cmt/preset';
 import { CodeModelConfiguration, CodeModelContent, CodeModelFileGroup, CodeModelProject, CodeModelTarget } from '@cmt/drivers/codeModel';
 import { ConfigureTrigger } from '@cmt/cmakeProject';
+import { onConfigureSettingsChange } from '@cmt/ui/util';
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
@@ -307,8 +308,9 @@ export class CMakeServerDriver extends CMakeDriver {
      * Track if the user changes the settings of the configure via settings.json
      */
     private _hadConfigurationChanged = true;
-    protected doConfigureSettingsChange() {
+    protected async doConfigureSettingsChange(): Promise<void> {
         this._hadConfigurationChanged = true;
+        await onConfigureSettingsChange();
     }
 
     async checkNeedsReconfigure(): Promise<boolean> {